+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## Create: Integrated Farming 1.1.3
+## Create: Integrated Farming 1.2.0
 
-Update for Create ***6.0.7***.
+Update for Create ***6.0.8***. Require Create: Dragons+ 1.8.4.
+
+### Update
+
+### Fix
+* Fix a rare client startup crash related to Ponder loading (#30)